docker/requirements.txt
python-telegram-bot~=22.1
python-dotenv~=1.1.0
ldap3~=2.9.1
tabulate~=0.9.0